Refactor wolrd api into separate module
