+++ api_world.py
@@ -1,0 +1,39 @@
+# For tutorial on FastAPI
+# See https://fastapi.tiangolo.com/tutorial/first-steps/
+from fastapi import APIRouter
+
+from utils import get_data_from_csv
+
+data_country = get_data_from_csv("world_table_country.csv")
+data_city = get_data_from_csv("world_table_city.csv")
+
+world_router = APIRouter()
+
+@world_router.get("/hello")
+async def say_hello() -> dict:
+    return {"message": "Hello!"}
+
+# define function that handles "GET" request with endpoint "/world" and "/world/"
+@world_router.get("/world")
+async def read_countries() -> dict:
+    return {"result": data_country}
+
+
+# define function that handles "GET" request with endpoint "/world/country/{name}"
+# "/world/country/{name}" is a "path paramter" endpoint
+@world_router.get("/world/country/{name}")
+async def read_country(name: str) -> dict:
+    for row in data_country:
+        if row["Name"].lower() == name.lower():
+            return {"result": row}
+    return {"result": {}}
+
+
+# define function that handles "GET" request with endpoint "/world/city/{name}"
+# "/world/city/{name}" is a "path paramter" endpoint
+@world_router.get("/world/city/{name}")
+async def read_city(name: str) -> dict:
+    for row in data_city:
+        if row["Name"].lower() == name.lower():
+            return {"result": row}
+    return {"result": {}}
